djTaskLine/dj_app/requirements.txt

Django

### PostgresSQL driver
psycopg2>=2.5.4

### MySQL driver
# mysqlclient

### Application server https://uwsgi-docs.readthedocs.io/en/latest/
uwsgi

### logging https://opbeat.com/
opbeat

### Auth
django-allauth

### http://www.django-rest-framework.org/
djangorestframework
markdown
django-filter
coreapi

##########################################
# dev and test -- must be removed for prod

django_extensions
django-debug-toolbar
jupyter


# test

factory_boy